--- batch/src/main/java/com/example/drrrbatch/batch/config/CurrentOSProperty.java
@@ -1,9 +1,0 @@
-package com.example.drrrbatch.batch.config;
-
-public class CurrentOSProperty {
-
-
-    static class WebDriverPath {
-
-    }
-}
