+++ batch/src/main/resources/application-batch.yml
@@ -25,8 +25,11 @@ logging:
 
 
 
-os:
-  mac:
-    webdriver: /usr/local/bin/chromedriver
+
 server:
   port: 0
+
+chromedriver:
+  path:
+    mac: /usr/local/bin/chromedriver
+    window: #윈도우 크롬드라이브 경로
